#84 - removed 'chosen answers' field from result #84 - removed unused exception #84 - created separated csv for results for one specified tournament

+++ e-arbiter-tournament-results/src/main/java/pl/cyganki/results/model/database/Result.java
@@ -36,9 +36,9 @@ public abstract class Result {
                 .sum();
     }
 
-    public List<Long> getListOfEarnedPoints() {
+    public List<Integer> getListOfEarnedPoints() {
         return Arrays.stream(this.earnedPoints.split(","))
-                .mapToLong(Long::parseLong)
+                .mapToInt(Integer::parseInt)
                 .boxed()
                 .collect(Collectors.toList());
     }
